move set title func to helper
