+++ src/components/forms/PlanForm/tests/index.test.tsx
@@ -7,14 +7,19 @@ import flushPromises from 'flush-promises';
 import { cloneDeep } from 'lodash';
 import React from 'react';
 import { MemoryRouter } from 'react-router-dom';
-import PlanForm, { defaultInitialValues, handlePlanTitleChange, propsForUpdatingPlans } from '..';
+import PlanForm, { defaultInitialValues, propsForUpdatingPlans } from '..';
 import { AN_ERROR_OCCURRED } from '../../../../configs/lang';
 import * as helperErrors from '../../../../helpers/errors';
 import { OpenSRPAPIResponse } from '../../../../services/opensrp/tests/fixtures/session';
 import store from '../../../../store';
 import { plans } from '../../../../store/ducks/opensrp/PlanDefinition/tests/fixtures';
 import { InterventionType, PlanStatus } from '../../../../store/ducks/plans';
-import { generatePlanDefinition, getPlanFormValues, planActivitiesMap } from '../helpers';
+import {
+  generatePlanDefinition,
+  getPlanFormValues,
+  handlePlanTitleChange,
+  planActivitiesMap,
+} from '../helpers';
 import * as fixtures from './fixtures';
 
 /* tslint:disable-next-line no-var-requires */
